+++ ci4/app/Libraries/DeploymentSteps/ClusterRoleBindingStep.php
@@ -43,7 +43,7 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
         return false;
     }
 
-    public function getSuccessStatus(): string {
+    public function getSuccessStatus(Deployment $deployment): string {
         return DeploymentStepHelper::ClusterRoleBinding_Found;
     }