+++ ci4/app/Libraries/DeploymentSteps/NamespaceStep.php
@@ -43,7 +43,7 @@ class NamespaceStep extends BaseDeploymentStep {
         return true;
     }
 
-    public function getSuccessStatus(): string {
+    public function getSuccessStatus(Deployment $deployment): string {
         return DeploymentStepHelper::Namespace_Found;
     }