+++ ci4/app/Libraries/DeploymentSteps/PersistentVolumeStep.php
@@ -43,7 +43,7 @@ class PersistentVolumeStep extends BaseDeploymentStep {
         return false;
     }
 
-    public function getSuccessStatus(): string {
+    public function getSuccessStatus(Deployment $deployment): string {
         return DeploymentStepHelper::PersistentVolume_Found;
     }