+++ ci4/app/Libraries/DeploymentSteps/ServiceAccountStep.php
@@ -42,7 +42,7 @@ class ServiceAccountStep extends BaseDeploymentStep {
         return false;
     }
 
-    public function getSuccessStatus(): string {
+    public function getSuccessStatus(Deployment $deployment): string {
         return DeploymentStepHelper::ServiceAccount_Found;
     }